scripts/requirements.txt
groq>=0.5.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
python-slugify>=8.0.0
